requirements.actions.txt
matplotlib
scrython
requests
bs4
tqdm
pandas
GitPython
python-dotenv